Replaced LED class with module and reworked hardware.h
A redesign of hardware.h was necessary to rework the LED code. 
Now only the code that is used is getting compiled (analog or neopixel - not both!) and there is no OOP overhead for the class.
Before, the LED class could have existed multiple times and needed an global instance.
Now the LED module exists only once by default and can be used everywhere without initialization or tossing around references.